+++ lib/ReactViewModels/ViewState.ts
@@ -1,30 +1,30 @@
+import {
+  action,
+  computed,
+  IReactionDisposer,
+  observable,
+  reaction,
+  runInAction
+} from "mobx";
 import { Ref } from "react";
 import clone from "terriajs-cesium/Source/Core/clone";
 import defined from "terriajs-cesium/Source/Core/defined";
-import DisclaimerHandler from "./DisclaimerHandler";
+import CesiumEvent from "terriajs-cesium/Source/Core/Event";
 import addedByUser from "../Core/addedByUser";
-import getAncestors from "../Models/getAncestors";
-import SearchState from "./SearchState";
-import Terria from "../Models/Terria";
+import isDefined from "../Core/isDefined";
 import triggerResize from "../Core/triggerResize";
-import {
-  observable,
-  reaction,
-  IReactionDisposer,
-  action,
-  runInAction,
-  computed
-} from "mobx";
-import { BaseModel } from "../Models/Model";
 import PickedFeatures from "../Map/PickedFeatures";
-import {
-  TourPoint,
-  defaultTourPoints,
-  RelativePosition
-} from "./defaultTourPoints";
-
+import getAncestors from "../Models/getAncestors";
+import { BaseModel } from "../Models/Model";
+import Terria from "../Models/Terria";
 import { SATELLITE_HELP_PROMPT_KEY } from "../ReactViews/HelpScreens/SatelliteHelpPrompt";
-import { LOCAL_PROPERTY_KEY as WELCOME_PROPERTY_KEY } from "../ReactViews/WelcomeMessage/WelcomeMessage";
+import {
+  defaultTourPoints,
+  RelativePosition,
+  TourPoint
+} from "./defaultTourPoints";
+import DisclaimerHandler from "./DisclaimerHandler";
+import SearchState from "./SearchState";
 
 export const DATA_CATALOG_NAME = "data-catalog";
 export const USER_DATA_NAME = "my-data";
@@ -38,6 +38,19 @@ interface ViewStateOptions {
   catalogSearchProvider: any;
   locationSearchProviders: any[];
   errorHandlingProvider?: any;
+}
+
+export interface Notification {
+  title: string;
+  message: string;
+  confirmText?: string;
+  denyText?: string;
+  confirmAction?: () => void;
+  denyAction?: () => void;
+  hideUi?: boolean;
+  type?: string;
+  width?: number | string;
+  height?: number | string;
 }
 
 /**
@@ -64,7 +77,7 @@ export default class ViewState {
   @observable isDraggingDroppingFile: boolean = false;
   @observable mobileView: string | null = null;
   @observable isMapFullScreen: boolean = false;
-  @observable readonly notifications: any[] = [];
+  @observable readonly notifications: Notification[] = [];
   @observable myDataIsUploadView: boolean = true;
   @observable mobileMenuVisible: boolean = false;
   @observable explorerPanelAnimating: boolean = false;
@@ -279,7 +292,8 @@ export default class ViewState {
    */
   @observable currentTool?: Tool;
 
-  private _unsubscribeErrorListener: any;
+  private _unsubscribeErrorListener: CesiumEvent.RemoveCallback;
+  private _unsubscribeNotificationListener: CesiumEvent.RemoveCallback;
   private _pickedFeaturesSubscription: IReactionDisposer;
   private _disclaimerVisibleSubscription: IReactionDisposer;
   private _isMapFullScreenSubscription: IReactionDisposer;
@@ -303,6 +317,23 @@ export default class ViewState {
       ? options.errorHandlingProvider
       : null;
     this.terria = terria;
+
+    this._unsubscribeNotificationListener = terria.notification.addEventListener(
+      (notification: Notification) => {
+        // Only add this notification if an identical one doesn't already exist.
+        if (
+          this.notifications.filter(
+            item =>
+              item.title === notification.title &&
+              item.message === notification.message
+          ).length === 0
+        ) {
+          runInAction(() => {
+            this.notifications.push(clone(notification));
+          });
+        }
+      }
+    );
 
     // Show errors to the user as notifications.
     this._unsubscribeErrorListener = terria.error.addEventListener(<any>((
@@ -445,6 +476,7 @@ export default class ViewState {
   dispose() {
     this._pickedFeaturesSubscription();
     this._disclaimerVisibleSubscription();
+    this._unsubscribeNotificationListener();
     this._unsubscribeErrorListener();
     this._mobileMenuSubscription();
     this._isMapFullScreenSubscription();
@@ -603,11 +635,14 @@ export default class ViewState {
   }
 
   getNextNotification() {
-    return this.notifications.length && this.notifications[0];
+    return this.notifications.length > 0 ? this.notifications[0] : undefined;
   }
 
   hideMapUi() {
-    return this.getNextNotification() && this.getNextNotification().hideUi;
+    return (
+      isDefined(this.getNextNotification()) &&
+      this.getNextNotification()!.hideUi
+    );
   }
 
   toggleFeaturePrompt(